requirements_updated.txt
flask
numpy
pandas
scikit-learn
xgboost
tensorflow
langchain
langchain_groq
python-dotenv
